Removed unused imports, preparing settings repository for future refactor
