+++ package.json
@@ -32,9 +32,7 @@
     "@types/react-dom": "^18.3.0",
     "framer-motion": "^11.2.3",
     "jest": "^29.7.0",
-    "parse-domain": "^8.0.2",
     "prettier": "^3.2.4",
-    "prettier-plugin-organize-imports": "^3.2.4",
     "react": "^18.3.1",
     "react-icons": "^5.2.1",
     "ts-jest": "^29.1.4",
