+++ src/data/models/BaseData.ts
@@ -1,0 +1,7 @@
+import { Category } from "./Category"
+
+type BaseTextData = {
+    text: string
+    type: Category
+    magnitude: number
+}
