+++ src/view/content/sections/OptionsSection.tsx
@@ -1,6 +1,7 @@
 import { Divider, Stack } from "@chakra-ui/react"
 import React from "react"
-import { Category, SiteData, SiteSeen } from "../../../data/models/SiteData"
+import { Category, SiteSeen } from "../../../data/models/Category"
+import { SiteData, } from "../../../data/models/SiteData"
 import {
 	addSiteUseCase,
 	reclassifySiteUseCase,
