+++ stories/ContentView.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react"
 import React from "react"
-import { Category } from "../src/data/models/SiteData"
+import { Category } from "../src/data/models/Category"
 import { ContentView } from "../src/view/content/ContentView"
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
